Move test case to LiveStripeResponseGetterTest
