+++ src/test/java/com/stripe/functional/LiveStripeResponseGetterTest.java
@@ -1,0 +1,44 @@
+import static org.hamcrest.MatcherAssert.assertThat;
+import static org.junit.jupiter.api.Assertions.assertNotNull;
+import static org.junit.jupiter.api.Assertions.assertThrows;
+
+import com.google.gson.JsonSyntaxException;
+import com.stripe.BaseStripeTest;
+import com.stripe.exception.ApiException;
+import com.stripe.exception.StripeException;
+import com.stripe.model.Subscription;
+import com.stripe.net.ApiResource;
+import com.stripe.net.HttpClient;
+import com.stripe.net.HttpHeaders;
+import com.stripe.net.HttpURLConnectionClient;
+import com.stripe.net.LiveStripeResponseGetter;
+import com.stripe.net.StripeRequest;
+import com.stripe.net.StripeResponse;
+import com.stripe.net.StripeResponseGetter;
+import java.util.Collections;
+import org.hamcrest.CoreMatchers;
+import org.junit.jupiter.api.Test;
+import org.mockito.Mockito;
+
+public class LiveStripeResponseGetterTest extends BaseStripeTest {
+
+  @Test
+  public void testInvalidJson() throws StripeException {
+    HttpClient spy = Mockito.spy(new HttpURLConnectionClient());
+    StripeResponseGetter srg = new LiveStripeResponseGetter(spy);
+    ApiResource.setStripeResponseGetter(srg);
+    StripeResponse response =
+        new StripeResponse(200, HttpHeaders.of(Collections.emptyMap()), "invalid JSON");
+    Mockito.doReturn(response).when(spy).requestWithRetries(Mockito.<StripeRequest>any());
+    Exception exception =
+        assertThrows(
+            ApiException.class,
+            () -> {
+              Subscription.retrieve("sub_123");
+            });
+    assertThat(
+        exception.getMessage(), CoreMatchers.containsString("Invalid response object from API"));
+    assertNotNull(exception.getCause());
+    assertThat(exception.getCause(), CoreMatchers.instanceOf(JsonSyntaxException.class));
+  }
+}
